requirements/frontend.txt
# SHA1:2b14b29ff0ab69c3e479920a2e2220cd7f7f53cb
#
# This file is autogenerated by pip-compile-multi
# To update, run:
#
#    pip-compile-multi
#
beautifulsoup4==4.12.3
    # via -r source/requirements.in
docutils==0.21.2
    # via -r source/requirements.in
jinja2==3.1.5
    # via -r source/requirements.in
markdown-it-py==3.0.0
    # via -r source/requirements.in
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
pillow==11.0.0
    # via -r source/requirements.in
soupsieve==2.6
    # via beautifulsoup4
